Make scope declarations stretchy buffers
The logic behind this is that the scopes really do need to
persist. Thus, although keeping them on the stack was nifty, it
required unnecessary copying and a second resolution.

The reason why this is especially useful will come in the next
commit, where we'll be able to store the declaration of an
expression every time we refer to an identifier. That will allow
resolution of values, keeping their declarations as they go.