Fix a nasty bug in declaration
identref can't really keep the decl where it was defined,
becasue the decls is a stretchy buffer, and thus holding on
to a reference to an element in this buffer (especially while it
is growing) is not OK. This isn't a problem elsewhere, because
typically the buffers aren't used until they are completely parsed
(e.g., formal parameters, record fields)

The "solution" to this problem is to have ident references store
their qualident, rather than the definition itself. This means
that the resolver needs to re-resolve the declarations, from their
names. It doesn't really do this correctly except for some really
simple cases, but it does kinda work.

It fixes a bug where the declaration was being corrupted, which I
was lucky to have caught by virtual of the buffer growing.

Also fixed a more subtle bug, where arrays were processed backwards