Get rid of fixed point for topological sort
The fixed point logic used before was incorrect, because if
function a depended on b and function c, and function c depended on
b, the order would be wrong.

The original version of this was the correct one, do a DFS by
starting with the main module, visiting every dependency, and *then*
adding to the fully resolved list. With this solution, the list
is now generated in the proper dependency order, no need for
reveral.

The original reason I didn't want to do this was because of scopes,
fearing that inner visited procedures could inadvertantly capture
outer ones.  This turns out not to be the case, because names are
properly resolved already.

This could be a problem with nested procedure scopes, but those
don't work right now anyway.