Fix open array assignments
If an open array is passed by value, it is readonly. This caused
Lex.Mod to not be quite right, so it was fixed.

However, passign open arrays as VAR parameters makes the code
generation messy, and, in C, unnecessary, since the parameter
(array) was already being passed by its pointer.

This means that there is a weird special case in the C code generator
that changes an open array to not be treated as such.