Pass all records as pointers.
This is an optimization that makes for much faster C code. Previously,
records (not marked with VAR) were passed by value, becuase that is
consistent with their defintion. However, in Oberon, this is unnecessary
because records passed as non-var parameters are read-only anyway.

Passing them by references is a big win, because some structures are
very large, like the Lexer, and this avoids unnecesary copies.