Initial support for imports
Imports are now included in the environment, and use an import cache to
keep a mappping from import to their decls. Those are stored in a
special import type.

The most complicated bit is a trampoline that goes like this

Module parsed
Finds import
Calls Lex.ImportHelper
Lexer state saved
Fetches the new module
Calls a callback in Parse
ImportHelper callback (in parse) is called
ParseModule for new module called
callback returns
Lex restores the state
the decls from the previous parseModule call are saved with the import