Move InitGlobalScope into ParseModule
Every module should be parsed from a pristine copy of the scope, which
points to the global decls (NIL, INTEGER, etc.) which should be shared
between all modules. This is especially important because ParseModule is
called recursively, through the import mechanism.

When an import is complete, the environment should be restored to the
exact state it was in. This involves saving and r4estoring the scope
state in the module.