Generate type tag initializers correctly
Whenever a structure is initialized, it *and any substructures* that
need typeinfo need that typeinfo initialized.  That is, if you have an
array of structs containing arrays of records, and those records need
typeinfo, then whenever the structure is created, *all* of the records
in the array need typeinfo initializatino.

Note this is broken for global variables for now, which do not get
typeinfo type ids.