Properly generate lengths for open arrays
Whenever an open array is passed as a parameter, the length of the open
array is passed as the next parameter. This is necessary in order to
implement the LEN function, which returns the length of an (open) array.
It could also form the basis of a bounds checking procedure, since it
can determine the length of any array, either with static type
information, or by looking at the passed in length param.

Though it would be better to only pass this around when it is actually
used, it does have the advantage that there is not per-structure bloat.
The array doesn't need to keep its length in a field, rather, the
complier can generate a length (or a reasonable reference to a length)
on demand.