Resolve CASE statements
CASE statements ended up re-using a great deal of code, and didn't end
up having a lot of necessary custom resoltion logic, by parsing as
if something was an if statement.

For example, the statement
  CASE a OF
    1, 2, 3 : ...
    4..6 : ...

passes "a" down to the individual items (to do type checking).  Thus,
this creates a structure like this
(CASE a
  (CASEITEM (= a 1) (= a 2) (= a 3) ...)
  (CASEITEM (tCaseDotDot (= a 4) (= a 6)) ...))

This will require a bit of work in the code generator to possibly break
apart these synthetic = operators, but in some code generation
scenarios, they can be valuable.