Ensure proper import order.
The module we wish to compile should be the top-level node that we want
to compile. All other nodes depend on it. Thus, we visit _module_init_
and then all of its dependencies, including imports.

The solution to this problem is to give every module an init method,
regardless of wheter or not it needs to do custom initialization. This
simplifies things later, as when a module is imported, the import node
will always be a _module_init_ method, followed by any of the
dependencies (imports) of that module.

This structure can be used in the visitor, which ensure that a module's
improts are visited before the module's init itself completes. Thus,
from a procedure dependency standpoint, a module's dependencies
(imports) will be processed first.