Visit types
Types are interesting because they can very easily be circular, with
pointers to other types. Thus, types cannot natually be added to the
list of visited declarations, but it can be ensured that only the types
that are used in a program are actually marked as visited, and thus need
to be generated by the code generator.