Generate code in four passes
Instead of just going through the declarations in order, the C code
generator first generates pointer declarations, then non-pointer type
declarations, then global var declarations, and then what's left over
(really, procedure declarations)

The reason for this is to make the order of declarations somewhat more
predictable and easier for a person to read. When comparing with the
eventaul oberon-generated code, the order should be much closer and
easier to follow.