Turn string constants into character constants
The AST resolver knows that the comparator type is char to char or
string to string. Now, if one of the arguments is a string, convert it
here to a char (using iVal for the character instead of the string. This
will make life easier in the code generator, which will get character
constants and can generate them appropriately.