Fix a very subtle bug in FOR loop direction
If either side of the FOR loop is not constant, it is unknown if the for
loop should be constant, and thus it should be forwards. Technically,
Oberon doesn't allow non constant for loops, but this is a nice
extension, and this should fix it.

With this change, the compiler passes the triple test!
  a) Use C compiler to generate out.prg, the oberon compiler
  b) use out.prg to generate an output compiler, the oberon code
     compiling itself, creating gen.c
  c) Compile gen.c, producing a new binary
  d) run the new binary. Its output should be the same as gen.c, which
     it is!