Major reorganization - out of dir builds
This commit makes it so that all generated files are created within the
build/ subdirectory, including resulting binaries.

In the process, the various parts of the compiler have moved around as
well.  The new directory structure is
  * c_bootstrap/ - files needed to build the C-based comiler
  * compiler/ - files needed to build the Oberon compiler
  * tests/ - example and test code

To support this, the c-based compiler inherited the concept of a build
path. It checks a series of directories in order to find source files,
which is important because its modules may not be in the same directory
as the c compiler itself.

The oberon compiler uses only the current directory. To support this,
the compile shell script copies any needed oberon files to the build/
directory, and then compiles them there.