More comprehensive tests of native values.
If an oberon structure contains only a single native record, e.g.

  RECORD _n :NATIVE INTEGER END;

Then, in the C code, the entire record can be cast to the native type,
as shown in the more comprehensive example.

This is perhaps not the best solution, as it might be good to have
native fields interspersed with oberon fields. Technically, this works,
but requires the C code (which currently is compiled beforehand) to
reference fields within the structure (which is defined in the code
generator).