Fix default case bug in parsing
If Parser.NextToken is called for an invalid character, the
compiler should report that illegal character. Before this, it would
instead loop forever, not parsing the character.

This was caught by trying to do something like Out.Char('A') which
would probably work in original Oberon (or C) but in Oberon-07,
character constants must be double quoted. This is debatable, but
at least now this correctly does not parse, and raises an error
when the parser encounters an invalid character. This would potentially
be true for any other character that appeared unexpectedly in the input
string, such as % which can't appear at the beginning of a lex string.