Temporarily zero out memory
This modifies the C code generator to zero out any block of memory
allocated with NEW. This paves the way for a code dumping utility where
all the fields are dumped, so the potentially invalid data is not
helpful for that dump.

This will be removed when memory clearing is no longer required.