Add debug information for types
This dumps type descriptions, caching in a similar way that declarations
themselves are cached. This should dump all reachable types for a
compilatino unit. It should also pave the way for some of the
machine-specific generation bits, like determining the size and
alignment of a type (e.g., size of an array, offsets of fields into a
record.)