Update field dump
This output is more readable, moving the fields to the end of a type
decl if they exist. This also asserts that RECORD fields end with a
declEND and that there is nothing past it. (This would not be true, for
example, for the procuedure var scope, which would have a declEND at the
end of the procedure vars, and then the next thing in the chain would be
the outer scope, for example.)