Add support for offsets
When generating code, many items have an "offset", like a global
variable (relative to the beginning of global data) or a field (relative
to the start of the field.)

This is only used in the code generator. Ideally, it would be in the
code generator directly. This would be possible, but since Oberon makes
it difficult to create a Map (in thie case, from declaration to offset)
it is currently added to the declaration itself, although only
referenced in the code generator.