Very early steps toward code generation
This introduces R5Compiler, which will output assembly code for a
hypothetical machine. The machine itself isn't well specified, and will
likely be quite similar to Wirth's RISC architecture (hence the R5
prefix).

Right now, the code generator calculates the sizes and offsets of global
and local variables, and generates function calls for the needed
functions, but that's about it. No code is generated (yet) for program
statements. However, type size counting is (hopefully) working.