Output LHS of assignment before RHS
Often, this doesn't matter, but with this change, my code matches
Wirth's exactly.

Wirth has a single pass compiler, so the code for the LHS MUST be
generated before the RHS. In my compiler, it isn't sensitive to order as
the AST exists in memory.

While the code, most of the time, should be semantically identical, it
is possible that there would be subtle differences. For example

a[SideEffectFunction(i)] := a[SideEffectFunction(i)];

If SideEffectFunction changes some global state and emits two different
values at different times, then the order of evaluation does indeed
matter.

With the exception of short circuit boolean evaluation, the order of
evaluation within an expression is (or probably should be) completely
unspecified, so evaluating the LHS or RHS of assignment should not
matter, just as F(a) + G(b) should not necessarily specify whether F or
G is evaluated first.

That said, I'll probalby keep this order, because it makes a bit more
sense.