Support pointer derferencing
When a pointer is loaded into a register, it is identical to an array
reference, because what is now in a register is a pointer (a memory
address). So, referenicing a pointer converts something to a register
indirect item.

Example:

MODULE S;

TYPE
  PList = POINTER TO LList;
  LList = RECORD
            i :INTEGER;
            next :PList;
            j :INTEGER
          END;

VAR
  i :INTEGER;
  p :PList;

BEGIN
  i := p.i;
  i := p.j;
  p.j := p.i;
  i := p.next.i;
  i := p.next.next.i + p.next.next.next.j
END S.

Yields

; S_i offset 0
; S_p offset 4
S__module_init_: code offset: 0
0	LDW R0, GB, #4 ; i := p.i;
1	LDW R0, R0, #0
2	STW R0, GB, #0
3	LDW R0, GB, #4 ; i := p.j;
4	LDW R0, R0, #8
5	STW R0, GB, #0
6	LDW R0, GB, #4 ; p.j := p.i;
7	LDW R1, GB, #4
8	LDW R1, R1, #0
9	STW R1, R0, #8
10	LDW R0, GB, #4 ; i := p.next.i;
11	LDW R0, R0, #4
12	LDW R0, R0, #0
13	STW R0, GB, #0
14	LDW R0, GB, #4 ; i := p.next.next.i + p.next.next.next.j
15	LDW R0, R0, #4
16	LDW R0, R0, #4
17	LDW R1, GB, #4
18	LDW R1, R1, #4
19	LDW R1, R1, #4
20	LDW R1, R1, #4
21	LDW R0, R0, #0
22	LDW R1, R1, #8
23	ADD R0, R0, R1
24	STW R0, GB, #0
25	RTS
__start__:
10000	JSR 0 (S__module_init_)
10001	HALT

Note that a placeholder for null pointer checks stil should be added.