Update C Codegen to support nested procedures.
In Oberon-07, nested procedures are legal, e.g

PROCEDURE Outer;
  PROCEDURE Inner;
  END Inner;
BEGIN (* Outer *)
  Inner
END Outer;

In Oberon, it is possible for inner to access local variables of outer
(full frames, like Pascal) but in Oberon-07, strictly speaking, this is
not legal. This makes nested procedures of extremely limited usefulness.
The two main uses of nested procedures is to declare a convenience short
name for a function (e.g., a nested inner procedure for Log() or
something, or for mutually recursive procedures. The latter (mutally
recursive procedures) are common in the compiler, but to work around
this, you can always declare pointers to procedures. For example, if A
and B call each other, you could have

MODULE Module;
VAR
  A :PROCEDURE;
  B :PROCEDURE;
PROCEDURE AImpl; END AImpl;
PROCEDURE BImpl; END BImpl;
BEGIN
  A := AImpl;
  B := BImpl;
END Module.

Which is what the compiler currently does.

However, supporting nested procedures is part of the language, so this
adds partial support for it. In the risc codegen, it should be pretty
easy to support.

The biggest limitation right now is that access to outer variables from
inner procedures is somewhat supported by the compiler, but it shouldn't
be (and it does not generate valid code.)  I personally think that
references to outer scopes ought to be valid (like in pascal) as it
makes it a little more like a closure. However, supporting this in C is
extremely unlikely, and supporting this in risc will require a frame
pointer, so it may not happen.