Initial support for conditionals (IF statements)
This is not working code yet, but paves the way for emitting code for
conditions, such as IF, REPEAT, and WHILE.  Right now, only IF
statements are generated.

The conditions are not yet generated, but there are placeholders, and
perhaps the trickiest bit (figuring out where the jumps should go to)
should be working. This works by using the instructions generated as a
"chain". When a forward jump is created, the code generator leaves a
placeholder (the current PC) and when the target is reached, the
backward code can be fixed to point to the correct address.

Example:

MODULE Simple;

VAR
  i :INTEGER;
BEGIN
  i := 0;
  IF i > 10 THEN
    i := 100;
    i := 100;
  ELSIF i > 11 THEN
    i := 200;
    i := 201;
  ELSIF i > 12 THEN
    i := 300;
    i := 301;
  ELSE
    i := 400;
    i := 401;
  END;
  i := 999;
  i := 1000
END Simple.

Yields

; Simple_i offset 0
Simple__module_init_: code offset: 0
0	MOV R0, #0
1	STW R0, GB, #0
2	COND True branch: 4
3	COND False branch: 9
4	MOV R0, #100
5	STW R0, GB, #0
6	MOV R0, #100
7	STW R0, GB, #0
8	B 29
9	COND True branch: 11
10	COND False branch: 16
11	MOV R0, #200
12	STW R0, GB, #0
13	MOV R0, #201
14	STW R0, GB, #0
15	B 29
16	COND True branch: 18
17	COND False branch: 23
18	MOV R0, #300
19	STW R0, GB, #0
20	MOV R0, #301
21	STW R0, GB, #0
22	B 29
23	COND True branch: 25
24	COND False branch: 29
25	MOV R0, #400
26	STW R0, GB, #0
27	MOV R0, #401
28	STW R0, GB, #0
29	MOV R0, #999
30	STW R0, GB, #0
31	MOV R0, #1000
32	STW R0, GB, #0
33	RTS
__start__:
10000	JSR 0 (Simple__module_init_)
10001	HALT