Allow IF to be the first statement
This fixes a bug because if the very first instruciton needs to be fixed
up, it would not have been generated properly.

This should never happen in practice, because it is impossible for the
first instruction for a procedure to be a conditional (there will always
be some comparison first, but until real code is generated, this makes
the chain linking more obvious.