Initial support for comparisons
This replaces the dummy code for comparisons with actual code that uses
condition codes to branch properly.

This utilizes Wirth's condition codes, which have the nice advantage
that they are supported in his Risc architecture, and also that the
inverse of each operation (= vs #, < vs >=, etc) is a simple
transformation. This ability to negate is important because IF a > b
THEN S1 END; S2 has the opposite comparison, CMP a, b; BLE S2 is often
used.

This uses the subtraction operator for comparisons, but since the
subtraction result is never used, this is fine.

Example:

MODULE Simple;

VAR
  i :INTEGER;
BEGIN
  IF i > 10 THEN
    i := 100;
    i := 101;
  ELSIF i >= 11 THEN
    i := 200;
    i := 201;
  ELSIF i < 12 THEN
    i := 300;
    i := 301;
  ELSIF i = 13 THEN
    i := 400;
  ELSIF i # 14 THEN
    i := 500;
  ELSE
    i := 600;
    i := 601;
  END;
  i := 999;
  i := 1000
END Simple.

yields

; Simple_i offset 0
Simple__module_init_: code offset: 0
0	LDW R0, GB, #0
1	SUB R0, R0, #10
2	BGT 8
3	MOV R0, #100
4	STW R0, GB, #0
5	MOV R0, #101
6	STW R0, GB, #0
7	B 40
8	LDW R0, GB, #0
9	SUB R0, R0, #11
10	BLT 16
11	MOV R0, #200
12	STW R0, GB, #0
13	MOV R0, #201
14	STW R0, GB, #0
15	B 40
16	LDW R0, GB, #0
17	SUB R0, R0, #12
18	BGE 24
19	MOV R0, #300
20	STW R0, GB, #0
21	MOV R0, #301
22	STW R0, GB, #0
23	B 40
24	LDW R0, GB, #0
25	SUB R0, R0, #13
26	BNE 30
27	MOV R0, #400
28	STW R0, GB, #0
29	B 40
30	LDW R0, GB, #0
31	SUB R0, R0, #14
32	BEQ 36
33	MOV R0, #500
34	STW R0, GB, #0
35	B 40
36	MOV R0, #600
37	STW R0, GB, #0
38	MOV R0, #601
39	STW R0, GB, #0
40	MOV R0, #999
41	STW R0, GB, #0
42	MOV R0, #1000
43	STW R0, GB, #0
44	RTS
__start__:
10000	JSR 0 (Simple__module_init_)
10001	HALT