Give inner procedures unique names
This is a dirty hack, but is necessayry because all C procedures need to
be generated at top level. This is a very simple solution to the
problem, but because inner procedures are extremely rare, this should
work.

With this change, the following compiles with the C compiler:

MODULE A;

VAR
  i :INTEGER;

PROCEDURE Outer1;
  PROCEDURE Inner;
  BEGIN
    i := i + 1
  END Inner;

BEGIN
  Inner
END Outer1;

PROCEDURE Outer2;
  PROCEDURE Inner;
  BEGIN
    i := i + 2
  END Inner;

BEGIN
  Inner
END Outer2;

BEGIN
  Outer1; Outer2;
END A.