Change stack layout
Previously, code was relative to FP, sort of. Local parameters were
negative offsets, and procedure parameters were positive offsets.

Instead, the compiler now assumes that the top of the stack points to
the return stack pointer. Immediately below are procedure parameters
(first parameter at the top), followed by local variables (first closest
to the top). Where this is heading is that procedures will expect their
paramters in registers, and deposit the return value in R0. It is the
responsibility of the called procedure to reserve space on the stack,
and then to store the parameters in that reserved space if it needs to
(the compiler doesn't do this now, but it will.)

Also included here is some soon-to-be-replaced code to call procedures,
mostly to make test code compiled. To that end, TestBed.ob combines
FieldTest and CondTest. The temporary code allows calling procedures
with no arguments as statements.