Preliminary support for procedure calls
Conceptually, calling a procedure is a straightforward process. The
calling convention is that parameters are stored in registers, the first
parameter in r0, next in r1, and so on.

Thus the code for procedure generation has changed:

* On entry to the procedure, the compiler reserves stack space for the
  parameters and local variables.
* The link register (return address) is put on the stack.
* Each parameter is copied from a register to the space reserved on the
  stack.
* On exit, the expression for return value is evaluated, and the result
  is in r0.
* The saved return address is put back in the link register.
* The stack space is deallocated.
* The code jumps to the contents of the link register, returning to the
  point that was called.

When calling a procedure, there is an analagous setup:

* Any currently in use registers are saved (more on this later).
* For each actual parameter, evaluate the expression and store in a
  register..
* The Branch and Link (BL) instruction is used to jump to the procedure
  entry point. The link bit deposits PC+1 in the link register (LR)
* When control is returned, the stack is in the same place as when the
  procedure was called.
* If there was a return value, it is in r0. If necessary, move to an
  unused register.
* The in use registers are restored from the stack.

The reason why registers are saved/restored using the stack is to
support nested procedure calls, e.g. f(1, g(x), 3). First, 1 is put in
r0, but then the expression for g(x) needs to be evaluated, and
ultimately end up in r1 for when f() is called (and then 3 will be put
in r2). To make this happen, the state needs to be preserved so that
each procedure call starts putting its values in the right place (e.g.,
putting x in r0 before calling g()). The biggest detail here is that if
the values are saved on the stack, the stack pointer is now different
than at procedure entry (so, if x was at SP+4, it's probably at SP+8
now).

The current code appears to work, and supports basic use cases (integer
parameters, and no VAR parameters yet.) It needs more stress test cases
to gain confidence.