Initial support for Parameters
Parameters are essentially variables that are passed by pointer. Since
the compiler already supports register indirect addresses (and pointers)
this should be a small change. The address (pointer) should be passed in
the register, and then derefernced before being used.

However, the TestBed code generation doesn't yet look correct, but close
enough to be a reasonable checkpoint.