Fix compiler crash
When either the C or RISC compiler encountered TestBed.ob in this form,
it would seg fault. Clearing memory after calling NEW fixes the problem,
so it appears to be an issue with unitialized fields.

That said, it was 100% possible to reproduce with the current version of
TestBed.ob, but very difficult in others, so it is some subtle issue in
the walking of reachable decls. So, this broken version is preserved in
case there are other subtle issues.