Add an interpreter for the RISC instruction set
This adds a new C program that takes RISC instructions and executes
them. It is minimally functional, but implements all of the instruction
set output from the compiler, and much of functionality of Wirth's RISC,
including the ReadInt, WriteInt, WriteChar, and WriteLn primitives.

The interface between the "virtual machine" and the compiler is a minor
tweak to the codegen output format. In addition to the previous assembly
language output, the compiler can now also emit something that can be
embedded direclty with the C runner, e.g.,
 {ADD, R0, R1, R2, 0},

which, inside the C code, is a literal for an Instruction.

This is not a reasonable coding for a real virtual machine, but makes
the output from the codgegen easy, and makes the generated C code
relatively easy to read.

The compiler and runner will be combined, so that the output of the
compiler can simply be #include-d in C, yielding a working program.
While this has been tested with IOTest, this initial version merly has a
countdown program written by hand.