Add a command line flag to dump code
RCompiler can now be invoked with -dumpcode, which will emit the program
in a format appripriate for embedding in the C virtual machine. This
creates PROGRAM and START_PC. Both are read by the interpreter. The
interpreter now merely #includes "risc_code.txt", so the program gets
embeded in the resulting C compiled binary.

risc_code.txt is taken directly from IOTest.ob.