Implement "traps"
Traps are an alternative to builtin functions. The idea of a NATIVE
function was introduced for the C code generator, which is that a
procedure has a definition, but does not generate code. In C, it is
expected that there will be C code to implement these procedure calls.

This was rather naturally extended to the RISC code generator by giving
each NATIVE procedure a (negative) offset, specified in its definition.
This causes the C code generator to ignore these, but the RISC code
generator will provide a branch to a negative PC value, e.g., BL -10 for
Out.Int. The interpreter can look at the jump target for BL
instructions, and execute native code if necessary.

It is assumed that every NATIVE procedure in a program has a unique
negative trap number, but the compiler doesn't enforce this. However,
since native procedures are rare, maintaining the list by hand should
hopefully not be prohibitive.

The IOTest is updated to use the same Out module, and now only relies on
a primitive for ReadInt, which could easily also be replaced.