Always generate LHS before RHS
It is important to generate registers in a strict atack order, which
implies that the left hand side of a binary expression *always* be
generated before the right hand side. This is important to avoid
problems like
  x + y[i]

  Before, this would say
    * LHS x's item is a var, no code need be genrated
    * RHS y[i] generates code, which places reserves R0 for RHS
      binary expression then loads LHS and then RHS
    * at this point, R0 contains y[i] and R1 contains x, so the compiler
      generates ADD R1, R1, R0

This can be a problem, because the compiler then thinks R1 is unused,
and a subsequent operation can trample this result.