Remove comparison assert
These ASSERT statements were based on the incorrect assumption that CMP
Ra, Rb was the same as SUB R0, Ra, Rb. Instead, it should be the same as
SUB Ra, Ra, Rb, and thus need not be the first register.

This fixes boolean paramters which are not the first argument to a
procedure.