Properly support open arrays
Open arrays must pass their length on the stack, immediately after the
pointer to the array itself. For a variable, this length is static
(coming from the variable's declaration). Otherwise, it is a parameter.
This length is used so that if you pass (a :ARRAY OF X) you can ask for
LEN(a) which, for parameters, needs to be available at runtime.

Adds a few tests that deal with arrays, and some rudimentary support for
strings. Out.Str could be implemented now (but isn't yet), and static
strings don't exist yet either.