Allow compile-time string constants
With this change, the RISC codegen can print "Hello, world!"

This takes advantage of the string pool, so critical inside the compiler
to create an interned set of used strings and identifiers.. For code
generation, all of the static strings used in the program are added to a
new string, which is placed in memory directly after all global
variables. That way, loading a static string is the same as loading
something relative to the global base.

In the runner, the generated code puts the contents of the string pool,
which is loaded into memory one byte at a time. (This probably only
works properly on little endian machines.)