Preliminary support for pointers to procedures
Because of the essentially Harvard architecture of the virtual machine
right now, a pointer to a procedure is merely an offset to an
instruction, generated during code generation. Thus, loading a pointer
to a procedure into a variable is merely MOVI, R0, #<procedure offset>
to get the pointer, and the BL R0 to call the procedure.

With this change, some simple cases for procedure pointers work
properly, like p(a, b) and procedures as a parameter to another
procedure. However, some more obscure cases don't work yet, like an
array of procedures. Perhaps more important is that a procedure in a
record might not work right, which may require a redesign of this.

This introduces a new virtual instruction, which is branch and link,
but to a register instead of a PC relative address. in a real machine
this would simply be a flag to the branch instruction. For simplicity,
it is another instruction.