Rework pointer to procedure support.
This fixes problems when generating pointers to procedures in complex
structures, like arrays. It resolves this by always generating the LHS
of a procedure call, most of the time this generates no code (just looks
up the procedure), but it works as well if the procedure is a dynamic
value.