Compile INC and DEC
To compile increment and decrement, the compiler does a primitive form
of tree transformation, turning INC(x) into x := x + 1. On the RISC
architecture, this generates reasonable code for the normal cases, such
as
  LDW R0, SP, #4
  ADDI R0, R0, #1
  STW R0, SP, #4

which is typical for RISC which doesn't have built-in increment
operators, as some CISC processors might have.

Code could be better, as INC(a[i][j][k]) could perform the address
lookup for the array element once, but that code is relatively rare.