Follow Oberon-07 spec for FOR loops
In the previous version, the code
  FOR i := 10 TO 0 DO
(when start and end are constant, and start > end) would act like
  FOR i := 10 TO 0 BY -1 DO

Since it is obvious that the intent is to count downward. However, in
the oberon specification, this is incorrect. The direction of the loop
should only be determined by the step (BY).

This behavior may return in the future, but for now it seemed better to
follow the language specification more closely, where the downward code
without a negative step would be a no-op.