Compile ASSERT statements
ASSERT(b) rougly coincied to IF ~b THEN HALT END

Builtin assert now compiles code similar to this. With this change,
FibFact passes its assertions and compiles and runs cleanly.

Placeholders are left for the other kinds of program traps, dereferncing
a NIL pointer and accessing an array index outside of the array.
These are supported in the RISC virtual machine, but not yet in the code
generator.

The machine has been tweaked a bit as well, to disassemble the last few
instructions run, to give more context in the case of a failure.