Add support for array bounds checking
Before an array index is multiplied by size to determine its offset,
the index is available, and should be >= 0 and < the array's size. This
can be accomplished with the carry flag (used for unsigned comparisons),
which is simulated in the RISC runner.  A bounds check therefore becomes
   <compute index into Rn>

   MOVI R(n+1) #array size, or
   LDW R(n+1) the __len for an open array

   CMP Rn+1, Rn
   BHI -2 (array trap)

This is an overhead of 3 instructions per array lookup, to ensure array
access are valid.