Support branch-and-link
All branch instructions (not just BL) can cause link to be set. This is
useful before calling a trap, because that way the PC is more useful.

So, if Register A is set on a branch instruction (it was previously
unused), it is used to indicate a link before the branch statement.