Add TypeExt disassembly from Project Oberon
This used project-norebo to recompile TypeExt.ob using Wirth's project
oberon compiler.

The motivation behind this is that Wirth has an intriguing technique for
handling type descriptor and type tags that I'm considering using in
this compiler. In Wirth's scheme, objects don't need a type descriptor
field, but instead rely on other conventions. This reduction in memory
overhead would probably be desirable.

But first, this file is included so that it can be studied better.