Fix negated boolean chains
Thanks to De Morgan, we have
 ~(a & b)

To be the same as
 ~a OR ~b

In the code examle, that means
  IF (~((a > 10) & (b > 20))) THEN WriteInt(1) ELSE WriteInt(0) END;

is really the same as
  IF (~(a > 10) OR ~(b > 20)) THEN WriteInt(1) ELSE WriteInt(0) END;

or
  IF ((a <= 10) OR (b <= 20)) THEN WriteInt(1) ELSE WriteInt(0) END;

Which is now what the code produces.

Previously, the boolean condition was correctly inverted for a single
clause, but not for a chain of clauses.

The output of TestBed.ob is now correct.