Simply (and hopefully fix) bounds checks
Bounds checks are now re-enabled, and hopefully working properly.
The original code for bounds checks was overly complicated, generating
the length for arrays two times, when one was sufficient. This is
because the code generation for bounds is relatively simple for open
arrays, which was the most complex case before. The length for an open
array is always simply the open array parameter plus four (next
parameter on the stack.)