Implement z flag
Really, all the branches in RISC should be based on flags (N, Z, V, C),
but for simplicity the runner hasn't used them, but instead relied on
CMP setting up for a branch. Each ALU operation should set these flags.
For now, only Z gets set, and it is used for BNE and BEQ. This paves the
way for set operations, such as x IN {1, 3}, which will translate to (1
<< x) & (1 << 1 | 1 << 3)  == 0, so the BEQ test will need to work even
if it isn't preceded by a CMP.