Fix generation of trap branches
Branches such as BEQ are usually relative to something in the function,
so if they are positive, they should be corrected to make an absolute
offset. Negative addresses (such as nil and bounds checks) should not be
rewritten.

This was generating correct code for the first function call, but not
subsequent ones, which is why it wasn't caught earlier.