Implement SET {a .. b}
Set ranges {a..b} implies all bits from a to b (inclusive) are set to
one. This is generated ineffieicntly in a loop in the C code, and
slightly better here. The logic is to generate the set {0..b}, or a set
with b ones. Then, generate another set {0..a-1} (this might be the
empty set). The set subtraction operator S1 & ~S2 causes all the bits in
S2 to be cleared.

For the record, Wirth's compiler generates even better code, especially
for the cases where a or b are constant (here, they both always get
loaded). Wirth's generated code looks like this:

  s is -4, i is -8, j is -12

  s := {0 .. j};
   6	 80E0000C	LDR  R0 SP      12
   7	 5100FFFE	MOV  R1  R0     -2
   8	 00110000	LSL  R0  R1  R0
   9	 5007FFFF	XOR  R0  R0     -1
  10	 5004FFFF	AND  R0  R0     -1
  11	 A0E00004	STR   R0 SP       4

  s := {3 .. j};
  12	 80E0000C	LDR  R0 SP      12
  13	 5100FFFE	MOV  R1  R0     -2
  14	 00110000	LSL  R0  R1  R0
  15	 5007FFFF	XOR  R0  R0     -1
  16	 5004FFF8	AND  R0  R0     -8
  17	 A0E00004	STR   R0 SP       4

  i := 3;
  18	 40000003	MOV  R0  R0      3
  19	 A0E00008	STR   R0 SP       8

  s := {i .. 9};
  20	 80E00008	LDR  R0 SP       8
  21	 5100FFFF	MOV  R1  R0     -1
  22	 00110000	LSL  R0  R1  R0
  23	 5100FC00	MOV  R1  R0  -1024
  24	 00050001	ANN  R0  R0  R1
  25	 A0E00004	STR   R0 SP       4

  s := {i .. j};
  26	 80E00008	LDR  R0 SP       8
  27	 5100FFFF	MOV  R1  R0     -1
  28	 00110000	LSL  R0  R1  R0
  29	 81E0000C	LDR  R1 SP      12
  30	 5200FFFE	MOV  R2  R0     -2
  31	 01210001	LSL  R1  R2  R1
  32	 00050001	ANN  R0  R0  R1
  33	 A0E00004	STR   R0 SP       4