Fix redundant array generation for bounds check
This is a no-op most of the time, but when an open array is passed as a
parameter from a prodcedure that takes it as an open array, the pointer
to the array was getting passed twice.

This was unnecessary because the array pointer always comes directly
befor4e the LengthOf computation.

This should fix StrTest.