Simple support for RECORD copy
Records may be assigned to one another, as long as the RHS of the
assignment is the same type or an extension of the source type.

This is currently handled by a naive memory copy, using the fact that
the structure is a fixed size. It iterates through the structure and
emits proper read and write calls.

For example, the output for RecCopy looks like this:

; RecCopy_r2 offset 0
; RecCopy_r1 offset 12
RecCopy_DoCopy: code offset: 0
	; aa1 offset 4
	; aa2 offset 8
0	SUB SP, SP, #12
1	STW LR, SP, #0
2	STW R0, SP, #4
3	STW R1, SP, #8
4	LDW R0, SP, #4
5	LDW R1, SP, #8
6	LDW R2, R1, #0
7	STW R2, R0, #0
8	LDW R2, R1, #4
9	STW R2, R0, #4
10	LDW R2, R1, #8
11	STW R2, R0, #8
12	LDW LR, SP, #0
13	ADD SP, SP, #12
14	B LR
RecCopy__module_init_: code offset: 15
15	SUB SP, SP, #4
16	STW LR, SP, #0
17	MOV R0, #10
18	STW R0, GB, #0
19	MOV R0, #20
20	STW R0, GB, #4
21	MOV R0, #30
22	STW R0, GB, #8
23	ADD R0, GB, #12
24	ADD R1, GB, #0
25	BL 0 (RecCopy_DoCopy)
26	LDW R0, GB, #0
27	STW R0, GB, #12
28	LDW R0, GB, #4
29	STW R0, GB, #16
30	LDW R0, GB, #8
31	STW R0, GB, #20
32	LDW R0, GB, #12
33	MOV R1, #0
34	STW R0, R1, #-1
35	LDW R0, GB, #16
36	MOV R1, #0
37	STW R0, R1, #-1
38	LDW R0, GB, #20
39	MOV R1, #0
40	STW R0, R1, #-1
41	MOV R0, #0
42	STW R0, R0, #-3
43	LDW LR, SP, #0
44	ADD SP, SP, #4
45	B LR
__start__: 46
46	BL 15 (RecCopy__module_init_)
47	HALT

This works well for small structures, as there are no branches to slow
down execution. However, for large structures this will likely fill up
the code buffer quickly.