Support copying of arbitrary sized records
The previous direct copy does not work well for large structures,
because a structure with N integers will have 2*N instructions (one
load, one store). This can lead to code bloat.

Instead, if a structure has more than 5 items (that is, size of 30) it
is copied instead in a loop, like this:

48	ADD R0, GB, #2016  ; Load Destination address to R0
49	ADD R1, GB, #0     ; Load Source address to R1
50	MOV R2, #2012      ; Size of the copy
51	LDW R3, R1, #0     ; Copy
52	STW R3, R0, #0
53	ADD R0, R0, #4     ; Increment pointers
54	ADD R1, R1, #4
55	SUB R2, R2, #4     ; And loop while there are more elements
56	BGT 51

This has the advantage that the code is compact - 9 instructions for a
copy of any size. It has the distinct disadvantage that it is
significantly slower. Direct copy has 2 instructuions per item, a load
and a store. For looping, this becomes 6 (the load and store, pointer
updates, and looping).

The choice of 30 bytes was somewhat arbitrary and can be tweaked later,
but seemed a good tradeoff that small structures will be copied directly
and only emit 15 instructions.